feat: the menu learns French, the ships get names, the music plays on
Build-215 feedback, five requests in one pass:

- SHIPS NAMED: Falcon, Viper, Phoenix, Ghost (user-approved). Phoenix is
  the hpp hull -- for a game called Reborn, the ship that came back from
  the dead could hardly be called anything else. The Custom status line
  now reads "Falcon - Red" instead of "Ship 1 - Red".

- SOLO LIVES join the icon + "x3" counter display. One display, one habit;
  the 2010 icon row is gone everywhere.

- MUSIC PLAYS ON between acts. All four acts cue the same UNREALPM.mp3 at
  different offsets, and every scene load stopped + restarted it. The
  engine now remembers which file the native player holds and only touches
  it on a REAL track change (title <-> game) or a music-less scene; the
  backgrounding path clears the memory so a fresh foreground re-inits.

- MENU RESTRUCTURED (user design): home is Game Solo / Game Multi /
  Others. Game Multi -> Local / Online / Back (new MENU_MULTI_MODE
  screen); the two entries left Others, and Tutorial moved from home into
  their slot. Matchmaking cancel/abort paths land on the new screen.

- FRENCH MENUS when the device language is French (Native_IsFrenchLanguage,
  NSLocale on iOS, 0 elsewhere). MENU_Tr maps each English label at its
  build site; unknown strings pass through, so a forgotten label shows
  English rather than nothing. The font atlas had real Latin-1 accent
  glyphs all along (rows 12-15) -- unreachable for 15 years because the
  renderer indexed it through a SIGNED char: 0xE9 went negative and the
  arithmetic shift produced a negative row. Indexing is unsigned now, so
  Cr\xE9dits can finally spell itself.

Rig 199/199; netcode untouched except the abort-menu target.

+++ engine/src/renderer.c
@@ -383,8 +383,12 @@ void SCR_ConvertTextToVertices(const char* string, float size, short ss_cooX, sh
 	for (i=0; i < stringLength; i++) 
 	{
 		//	Log_Printf("character %c=%d\n",*currentChar,*currentChar);
-		textCoo[X] = *currentChar & 15;
-		textCoo[Y] = *currentChar >> 4;
+		// v2: UNSIGNED indexing -- the atlas has real Latin-1 accent glyphs
+		// (rows 12-15), but char is SIGNED on Apple ARM64: an accented byte
+		// (0xE9, 'e acute') went negative and the arithmetic shift produced
+		// a negative row. French needs the top half of the table.
+		textCoo[X] = (uchar)*currentChar & 15;
+		textCoo[Y] = (uchar)*currentChar >> 4;
 		
 		scr_p_TextVertices[0].pos[X] = ss_cooX-charWidth;
 		scr_p_TextVertices[0].pos[Y] = ss_cooY+charHeight;
